backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3

# Hardware communication dependencies
pyserial==3.5
toml==0.10.2

# Future hardware-specific dependencies (to be installed when SDKs are available)
# pyvisa==1.14.1  # For SCPI/VISA instrument communication
# numpy==1.24.3   # For data processing
# scipy==1.10.1   # For signal processing

